third_party/blink/web_tests/platform/mac-mac12/external/wpt/css/css-scroll-snap/input/snap-area-overflow-boundary-viewport-covering.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] Mouse-wheel scrolling with vertical snap-area overflow
  assert_equals: Wheel-scroll moved scroll position expected 50 but got 0
Harness: the test ran to completion.
